move jndi code to jndi maven module
