+++ jndi/src/main/java/org/huberb/ee8sample/jndi/impl/hierarchysearch/HierParser.java
@@ -30,7 +30,7 @@
  * maintenance of any nuclear facility. Licensee represents and warrants
  * that it will not use or redistribute the Software for such purposes. 
  */
-package org.huberb.ee8sample.fs.jndi.hierarchy;
+package org.huberb.ee8sample.jndi.impl.hierarchysearch;
 
 import java.util.Properties;
 import javax.naming.CompoundName;
