+++ jndi/src/test/java/org/huberb/ee8sample/jndi/impl/flat/FlatNameParserTest.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.huberb.ee8sample.fs.jndi.flat;
+package org.huberb.ee8sample.jndi.impl.flat;
 
+import org.huberb.ee8sample.jndi.impl.flat.FlatNameParser;
 import javax.naming.Name;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
